toml: Fixed empty table array index
Previously, an table array root entry would always start with array metakey #0, ignoring the possibility of an emptry table array. An empty table array would have the index 0, but no child keys with index 0. Now the array metakey will be set on the first emission of child (after a key/value, a simple table, or an inline comment on said table array).