toml: Fixed integer range checking
Previously, did not check any range for integers on reading/writing.
Now added checks for that. On reading, will emit an error if a syntactically correct
integer exceeds the range of ulonglong/longlong. On writing, value will
not be seen as integer, but written as string instead.

Also fixed incorrect float checking in type.c (Out-of-range integers
would get interpreted as floats, because commas were optional).